[P2P-54] Place, PlaceService 관련 테스트 기능 추가 등

+++ src/main/java/com/prgrms/p2p/domain/place/dto/DetailPlaceResponse.java
@@ -21,8 +21,8 @@ public class DetailPlaceResponse {
   private String phone;
   private String imageUrl;
 
-  private Long likeCount;
-  private Long usedCount;
+  private Integer likeCount;
+  private Integer usedCount;
 
   // TODO: 2022/07/29 사용된 코스 추천 몇가지 들어가야 함
 }
